requirements.txt
Pillow>=10.0.0
img2pdf>=0.5.0
streamlit>=1.28.0
